pyproject.toml
[project]
name = "autofix-ia"
version = "0.1.0"
description = ""
authors = [
    {name = "Gonzalo Ayuso",email = "gonzalo123@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-json-logger (>=4.0.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "strands-agents (>=1.22.0,<2.0.0)",
    "gitpython (>=3.1.46,<4.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "claude-code-sdk (>=0.0.25,<0.0.26)",
]

[tool.poetry]
packages = [{include = "autofix"}]


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
